ts compiler: simplify vm golden tests
