+++ compiler/x/ts/helpers_test.go
@@ -1,0 +1,18 @@
+package tscode_test
+
+import (
+	"bytes"
+)
+
+// stripHeader removes the generated header from compiler output so
+// golden comparisons remain stable.
+func stripHeader(b []byte) []byte {
+	if i := bytes.IndexByte(b, '\n'); i != -1 && bytes.HasPrefix(b, []byte("// Generated by Mochi")) {
+		b = b[i+1:]
+		if j := bytes.IndexByte(b, '\n'); j != -1 && bytes.HasPrefix(b, []byte("// Source:")) {
+			return b[j+1:]
+		}
+		return b
+	}
+	return b
+}
